1.发送微博导航条。2发送微博textview添加placeholder。3发送微博工具条，4.发送文字微博和图片微博

+++ CsWeibo/CsWeibo/Classes/Home/Controller/OHHomeViewController.m
@@ -293,7 +293,7 @@
  *
  *  @param count 最新微博数
  */
-- (void)showNewStatusCount:(int)count
+- (void)showNewStatusCount:(NSUInteger)count
 {
     UILabel *label = [[UILabel alloc] init];
     label.backgroundColor = [UIColor colorWithPatternImage:[UIImage imageNamed:@"timeline_new_status_background"]];
