+++ CsWeibo/CsWeibo/Classes/Home/View/OHStatusPhotosView.h
@@ -17,6 +17,6 @@
 /**
  *  根据图片个数计算相册的尺寸
  */
-+ (CGSize)sizeWithCount:(int)count;
++ (CGSize)sizeWithCount:(NSUInteger)count;
 
 @end
