+++ CsWeibo/CsWeibo/Classes/Home/View/OHStatusToolbar.m
@@ -95,7 +95,7 @@
 - (void)layoutSubviews{
     [super layoutSubviews];
     //布局
-    int btnCount = self.btns.count;
+    NSUInteger btnCount = self.btns.count;
     CGFloat btnW = self.width / btnCount;
     CGFloat btnH = self.height;
     for (int i = 0; i < btnCount; i++) {
@@ -107,7 +107,7 @@
     }
     
     //设置分割线
-    int dividerCount = self.dividers.count;
+    NSUInteger dividerCount = self.dividers.count;
     for (int i = 0; i<dividerCount; i++) {
         UIImageView *divider = self.dividers[i];
         divider.width = 1;
